とりあえずbuiltin の実行は完了。　しかしbuiltin以外の処理が無理
