+++ executor/test_executor.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 run() {
-cc  ../parser/*.c ../lexer/*.c ../linerlizer/*.c ./*.c ../builtin/*.c ../utils/*.c ../libft/libft.a 
+cc  -lreadline ../expander/*.c ../parser/*.c ../lexer/*.c ../linerlizer/*.c ./*.c ../builtin/*.c ../utils/*.c ../libft/libft.a 
 	input="$1"
 	# leak_check=(valgrind -q --leak-check=full)
 	#if you want to check leaking, add this command "${leak_check[0]}" to head of ↓
